+++ src/components/table/examples/table-custom-components.tsx
@@ -6,6 +6,25 @@ import { capitalize } from 'lodash-es';
 /**
  * Custom components
  *
+ * You can specify a custom component to use for any column in your table. This
+ * is done under the `component` key in the schema, following the
+ * [TableComponentDefinition](#/type/TableComponentDefinition/) specification,
+ * for example:
+ *
+ * ```ts
+ * const columns = [
+ *     {
+ *         title: 'Food',
+ *         field: 'food',
+ *         component: { name: 'my-fancy-food-displayer' },
+ *     },
+ * ];
+ * ```
+ *
+ * While you can, in principle, use any component in a table, your custom table
+ * components should implement the [TableComponent](#/type/TableComponent/)
+ * interface.
+ *
  * @link birds.ts
  * @link table-food.tsx
  */
@@ -37,7 +56,11 @@ export class TableExampleCustomComponents {
                 formatter: this.capitalizeStrings,
             },
             { title: 'Nest type', field: 'nest', formatter: capitalize },
-            { title: 'Eggs per clutch', field: 'eggs' },
+            {
+                title: 'Eggs per clutch',
+                field: 'eggs',
+                horizontalAlign: 'right',
+            },
             { title: 'Origin', field: 'origin' },
         ];
     }